add scroll listener

+++ app/src/main/java/project/celine/infinitescroll/data/RecordEvent.java
@@ -8,18 +8,18 @@ import project.celine.infinitescroll.model.RecordEntity;
  * Created by celine on 2015/10/26.
  */
 public class RecordEvent {
-    public long getStart() {
+    public int getStart() {
         return start;
     }
 
-    long start;
+    int start;
     public List<RecordEntity> getRecordEntities() {
         return recordEntities;
     }
 
     List<RecordEntity> recordEntities;
 
-    public RecordEvent(long start, List<RecordEntity> recordEntities) {
+    public RecordEvent(int start, List<RecordEntity> recordEntities) {
         this.recordEntities = recordEntities;
         this.start = start;
     }
